requirements.txt
asgiref==3.5.2
autopep8==1.6.0
certifi==2022.6.15
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.6
django-filter==22.1
django-rest-swagger==2.2.0
django-silk==5.0.1
djangorestframework==3.13.1
docutils==0.19
drf-yasg==1.21.0
gprof2dot==2021.2.21
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
lockfile==0.12.2
Markdown==3.3.7
MarkupSafe==2.1.1
openapi-codec==1.3.2
packaging==21.3
pid==3.0.4
Pillow==9.2.0
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyparsing==3.0.9
python-daemon==2.3.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
service==0.6.0
setproctitle==1.2.3
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.10
